requirements.txt
instabot==0.26.0
python-dotenv==0.10.1